とりあえず動くところまで書いた

+++ module/core/src/main/scala/jozepoko/stock_trader/core/domain/entity/FiveMinutelyStockPrice.scala
@@ -8,8 +8,8 @@ case class FiveMinutelyStockPrice(
   code: Int,
   market: Market,
   name: String,
-  openingPrice: Int,
-  highPrice: Int,
-  lowPrice: Int,
-  closingPrice: Int
+  openingPrice: Double,
+  highPrice: Double,
+  lowPrice: Double,
+  closingPrice: Double
 ) extends StockPrice
